preprocessors/requirements.txt
spacy
newspaper3k
opencv-python
mtcnn
numpy
Pillow
librosa
soundfile
